+++ lib/models/derived_linking_key.dart
@@ -1,0 +1,11 @@
+import 'dart:typed_data';
+
+class DerivedLinkingKey {
+  final BigInt privateKey;
+  final Uint8List compressedPublicKey;
+
+  DerivedLinkingKey({
+    required this.privateKey,
+    required this.compressedPublicKey,
+  });
+}
